发布版本V0.1

+++ src/main/java/com/masou/coupon/data/mappers/TicketTypeMapper.java
@@ -13,7 +13,7 @@ public interface TicketTypeMapper {
 
     int insertSelective(TicketType record);
 
-    TicketType selectByPrimaryKey(Long id);
+    TicketType selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(TicketType record);
